api_service/pyproject.toml
[tool.poetry]
name = "api_service"
version = "0.1.0"
description = ""
authors = ["gurland <stas.gurland@gmail.com>"]
license = "MIT"

[[tool.poetry.source]]
name = "test-pypi"
url = "https://test.pypi.org/simple"
secondary = true

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.5.0"
peewee = "^3.15.3"
connexion = { version = "^2.14.1", extras = ["swagger-ui"] }
psycopg2-binary = {version = "^2.9.3", source = "test-pypi"}
Flask-Cors = "^3.0.10"
redis = "^4.3.4"
uWSGI = "^2.0.20"
bcrypt = "^4.0.0"
PyJWT = "^2.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
